\name{human_pp}
\docType{data}
\alias{human_pp}
\title{Sample Human Protein Interactome}
\description{
A set of 864 human proteins involved in 999 interactions.
}
\usage{
human_pp
}
\format{Data frame with 2 columns.}
\source{\url{http://tagc.univ-mrs.fr/welcome/spip.php?rubrique197}}
\references{
Becker, E., Robisson, B., Chapple, C.E., Guenoche, A. and Brun, C. (2012) Multifunctional proteins revealed by overlapping clustering in protein interaction network. \emph{Bioinformatics} \bold{28}, 84-90.
}
\keyword{datasets}

