\name{karate}
\docType{data}
\alias{karate}
\title{Social Network in a Karate Club}
\description{
A social network of friendships between 34 members of a karate club at a US university in the 1970s (Zachary 1977).
}
\usage{
karate
}
\format{Data frame with 2 columns.}
\source{\url{https://github.com/gephi/gephi/wiki/Datasets}}
\references{
Zachary, W. W. (1977). An information flow model for conflict and fission in small groups. \emph{Journal of Anthropological Research} \bold{33}, 452-473.
}
\keyword{datasets}

