\name{lesmiserables}
\docType{data}
\alias{lesmiserables}
\title{Co-Appearance Network from Les Miserables}
\description{
The co-appearance network for Les Miserables (Knuth 1993). Involves 252 interactions among 77 nodes.
}
\usage{
lesmiserables
}
\format{Data frame with 2 columns.}
\source{\url{https://github.com/gephi/gephi/wiki/Datasets}}
\references{
Knuth, D. E. (1993). \emph{The Stanford GraphBase: A Platform for Combinatorial Computing}, Addison-Wesley, Reading, MA.
}
\keyword{datasets}

