\name{pp_rnapol}
\docType{data}
\alias{pp_rnapol}
\title{Sample Yeast Protein Interactome}
\description{
A set of 56 yeast proteins involved in 651 interactions related to transcription (Yu et al. 2008).
}
\usage{
pp_rnapol
}
\format{Data frame with 2 columns.}
\source{\url{http://interactome.dfci.harvard.edu}}
\references{
Yu, H., \emph{et al}. (2008). High-quality binary protein interaction map of the yeast interactome network. \emph{Science} \bold{322}, 104-110.
}
\keyword{datasets}

