\name{weighted}
\docType{data}
\alias{weighted}
\title{Sample Gene Co-Expression Network}
\description{
A sample of 200 links from a Drosophila gene co-expression network illustrating the input required for a weighted network (Tomancak et al. 2007).
}
\usage{
weighted
}
\format{Data frame with 3 columns.}
\source{\url{http://www.fruitfly.org/cgi-bin/ex/insitu.pl}}
\references{
Tomancak, P. \emph{et al}. (2007). Global analysis of patterns of gene expression during \emph{Drosophila} embryogenesis. \emph{Genome Biol} \bold{8}, 145.1-145.34.
}
\keyword{datasets}

